os/bluestore: try to gift large extents, then fall back to small extents
First try to gift using the larger alloc_size (normally bluefs_alloc_size,
but here max(bluefs_alloc_size,bluefs_shared_alloc_size) just in case the
settings are weird.  If that fails, then try the shared_alloc_size.  If
that fails, fail and complain as before, with an more accurate error
message.

This is a slightly modified version of the luminous patch,
91a1fa17470b158ed1fe22a725cd42ac28344237

Signed-off-by: Sage Weil <sage@redhat.com>