mgr/dashboard: CephFS directory component
Now it's possible to list directories of a CephFS in a tree view inside
the dashboard.

Currently the tree is not refreshed every 5 seconds as the listing of
directories is a heavier operation.

When a directory is clicked, the dashboard will fetch it's
subdirectories and the subdirectories of it's subdirectories in order to
show if the subdirectories have subdirectories.

The details of a directory show the full path, it's quota values and
the path they originate from and a list of all snapshots.

Fixes: https://tracker.ceph.com/issues/41575
Signed-off-by: Stephan Müller <smueller@suse.com>