rgw/rgw_rados.cc: store small object's data part into xattr
When the data size of the object is smaller than the minimum space
allocation unit on the osd side, a lot of disk space is wasted. If
store the data part of these small objects into xattr, the db of osd
will convert these small data into continuous large blocks, thus
reducing the waste of disk space.

Fixes: https://tracker.ceph.com/issues/41417
Signed-off-by: Yang Honggang <yanghonggang@umcloud.com>