rgw: Fix config parsing order and output
Currently, the profile parser properly interprets the s3 settings,
but it doesn't output them to the rgw log after parsing them.

Additionally, the "default" conf is parsed too early, preventing
the use of connection_id or acl_profiles. Because the explicit
profiles inherit the default profile, you can't have a connection{}
block in the default profile and connection_id in the explicit
profile. This allows the default profile to use connection_id, thus
allowing explicit profiles to supply a different connection_id.

Fixes: https://tracker.ceph.com/issues/41540
Signed-off-by: Edward Fisher <ed@debacle.org>