rgw: Fix multipart upload tracking+part size calculation
The minimum part size needs to be 1 byte bigger than obj_size/10000,
otherwise we end up with 10001 parts.

Additionally, there's no reason to track cur_ofs in the status
object, since it's safer to calculate it from cur_part and part_size.
Tracking cur_ofs separately can (and does) result in situations where
a part fails and cur_ofs is incremented but cur_part is not. This
eventually can cause an underflow in cur_part_info.size, resulting in
an incorrect content-length header and an EntityTooLarge error from
the remote endpoint.

Fixes: https://tracker.ceph.com/issues/41540
Signed-off-by: Edward Fisher <ed@debacle.org>