mgr/dashboard: Add styling for ui blocking notification
UI blocking notifications should only appear if an error occurs which makes the
frontend unusable. The UI will be released as soon as the error is fixed. For
example the UI would be blocked if the api is unreachable.

Signed-off-by: Sebastian Krah <skrah@suse.com>
(cherry picked from commit 662e17b05a3612ec260aedbf3ab60089b082010b)