mgr/dashboard: Handle always-on Ceph Manager modules correctly
Currently always-on modules are not marked as enabled in the WebUI and can be disabled. This PR will fix that.
Note, this PR will NOT implement code that will prevent a developer from trying to disable an always-on module through the REST API. The Mgr Python extension will throw an adequate exception.

This PR will also do:
* Remove old code fragments from a previous Mgr Module management UI that is obsolete now.
* Cleanup code in BaseMgrModule code.

Fixes: https://tracker.ceph.com/issues/41648

Signed-off-by: Volker Theile <vtheile@suse.com>
(cherry picked from commit 8ac655963755820c9bab3953a4195fbc916f7072)