mgr/dashboard: Improve images e2e test
The 'should move the image to the trash' e2e fails because the datatable content has not been updated at the time the test checks for the existence of the deleted image.

To fix that the helper function waits until the presence of the datatable row (the same approach is done by other helper functions, too).

Signed-off-by: Volker Theile <vtheile@suse.com>