auth: also catch end_of_buffer while parsing text input
[On FreeBSD]
in KeyRing::decode() text parsing in only executed if decode throws.
This results in an ABORT signal in bin/ceph when running vstart.sh
```
12: Populating config ...
12: /usr/home/Ceph/master/ceph/src/vstart.sh: line 1154: 70347 Done                    cat <<EOF
12: [global]
12: osd_pool_default_size = $OSD_POOL_DEFAULT_SIZE
12: osd_pool_default_min_size = 1
12: mon_pg_warn_min_per_osd = 3
...
...
12: [mgr]
12: mgr/telemetry/nag = false
12: mgr/telemetry/enable = false
12:
12: EOF
12:
12:      70348 Abort trap              (core dumped) | $CEPH_BIN/ceph -c $conf_fn config assimilate-conf -i -
```

Only buffer::error is caught in `KeyRing.cc`, but reading config text sometimes
lets the buffer_length run into absurd numbers. This results in a `copy()` call
with too large length.  And this generates `end_of_buffer`.

So we need to catch that as well.

However adding buffer::end_of_buffer as exception in another case
does not seem to work, because it is not matched????
So catching all exceptions, and continue to perhaps trap in  decode_plaintext()

fixes: https://tracker.ceph.com/issues/41717
Signed-off-by: Willem Jan Withagen <wjw@digiware.nl>