mon: use variadic templates for command classes
This allows us more flexibility when defining the command classes for
each service. For instance, some services (like the osdmon and the
monmapmon) will have both a pending and a stable state that is very well
defined (i.e., via incremental maps or whatever); other services, like
the authmon, do not, and using the existing templates was not an option.

By introducing variadic templates, we can have the services themselves
specifying the parameters that their commands will be taking (if any),
and thus allowing services like the authmon to implement these wrappers
in a friendlier, saner way.

The obvious drawback is that, at its current form, this is far from
obvious, and will need quite a bit of documenting. Hopefully, future
contributors will also be able to draw the blueprints for command
implementation for new services from what we are now implementing.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>