cls/rbd: expand MirrorPeer to annotate bidirectional relationships
The legacy MirrorPeer format is a RX direction and still requires
a client name for connecting to the remote peer. The cluster name
has been renamed internally to site name to match the new bootstrap
helper. The TX mode peers will be automatically created by remote
rbd-mirror instances that ping the local cluster.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>