crimson/osd: pass octopus to PeeringCtx's constructor
to accommodate the change of c290f9d5

classic OSD started to use single-pg peering ops since 68862636, and it
switched over to these ops since octopus. and it's assumed that
crimson-osd won't be GA until octopus+2 release, so it's not needed to
support pre-octopus releases. hence "octopus" is passed to PeeringCtx's
constructor.

Signed-off-by: Kefu Chai <kchai@redhat.com>