script/run_tox.sh: use WITH_PYTHON{2,3} to ensure python is installed
before this change, run_tox.sh is hardwired to python2.7,
`setup-virtualenv.sh` use it as the default python version. after this
change, we use python3 if WITH_PYTHON3 is enabled, and fallback to
python2 otherwise.

this is not quite right, as WITH_PYTHON2 and WITH_PYTHON3 are used to
specify what python version our python binding should be built. it's not
directly related to what python we should use for creating venv which is
in turn used for running tox based test.

but we can use these settings to ensure that the python interpreter is
available for creating the venv environment. normally, the venv is
prepared by `${name}-venv` target if one runs "make tests" for preparing
the dependencies for running "ctest" or "make check". but we should
allow user to use `run_tox.sh` directly for performing tox tests without
running "make tests" first.

Signed-off-by: Kefu Chai <kchai@redhat.com>