mds: change MDCache::discover_path()'s wants_xlocked semantic
If an inode's snaplock is locked by a MDRequest, the inode can not be
renamed by other MDRequest. For rename case, snaplocks of all inodes
in paths are locked. So the paths are stable while handling slave rename.
It's OK to discover all components (even they are xlocked) in the paths.

Signed-off-by: "Yan, Zheng" <zyan@redhat.com>