rgw: add `rgw-orphan-list` tool & `radosgw-admin bucket radoslist ...`
Adds a `radosgw-admin` subcommand and walks the associated bucket
indices and manifests to generate the list of rados objects that
represent the rgw objects in the bucket(s).

Also adds a tool named `rgw-orphan-list`, which uses the radoslist
subcommand, that produces a list in a local file of what appear to be
rgw orphans.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit e396064d9afa6fcba7b6bb1e3a500ef9e718e0d2)