rgw: add `rgw-orphan-list` tool & `radosgw-admin bucket radoslist ...`
Adds a radosgw-admin subcommand and possibly a bucket name (if not,
all buckets are assumed) and walks the bucket(s) listing(s) and the
manifest (if it exists) for each object in the listing to generate the
rados objects that represent the rgw objects in the bucket.

Also adds a tool named rgw-orphan-list that will produce a list in a
local file of what appear to be rgw orphans.

NOTE: This is not a cherry-pick from master because the feature was
originally written for an older version of Ceph and is being gradually
forward-ported to newer and newer versions.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>