cls/rbd: preliminary support for mirror image status from multiple sites
The mirror_image_status_set method will now accept a MirrorImageSiteStatus
structure. When no site name is provided, it is backwards compatible
with the original MirorrImageStatus parameter, and when a site name
is provided, the setter will fail if not using a new OSD.

The mirror_image_status_get/list methods will now return an extended
MirrorImageStatus which always includes the local site status and
optionally includes the remote site status.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>