mgr/k8sevents: Add support for remote kubernetes
The initial implementation sent events to the
kubernetes cluster Ceph is running under managed
by rook-ceph. This patch extends this support to
include sending events to an external kubernetes
cluster, that may just be consuming ceph resources

Additional docs added to help people use the
module, either with or without rook-ceph.

It also address a bug where the module was loaded
outside of a rook-ceph environment.

Fixes: https://tracker.ceph.com/issues/41737

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>