librados: add a RETURNVEC flag
This signals that the client is soliciting and expecting return codes
that are >0 and/or return buffers on update ops.

This should be used with caution, since it means that compound requests
that include read and write requests will *also* persist the read ops in
order to be idempotent, which means that those ops' results may get
truncated.

Signed-off-by: Sage Weil <sage@redhat.com>