librados: add a RETURNVEC flag
This signals that the client is soliciting and expecting return codes
that are >0 and/or return buffers on update ops.

This should be used with caution, since it means that compound requests
that include read and write requests will *also* persist the read ops in
order to be idempotent, which means that those ops' results may get
truncated.

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 330f6f3ae986aaf3ef4c4a5f2d2099da2a2f5d02)
Signed-off-by: Adam C. Emerson <aemerson@redhat.com>