ceph-volume: Volume.filter should not purge the object
Volume.filter removes items from object that does not match the filter.
This approach is problematic since the code calling this method has to
pass a copy of list of LVs, in case it needs it later. The better
approach would be that this method should return a new list/object
containing only the LVs that matches the filters.

Fixes: https://tracker.ceph.com/issues/42067
Signed-off-by: Rishabh Dave <ridave@redhat.com>