doc/releases: access main releases page from top-level TOC
Before this commit, one had to go through the releases index to get to the main
releases page. This should be the other way around.

Also, the main releases page should not be called "schedule.rst" - while it
contains a release timeline, that's not its principal purpose. Rename the page
to "general.rst" and move the "Understanding the release cycle" section before
the timeline.

Signed-off-by: Nathan Cutler <ncutler@suse.com>