qa/rados/upgrade/jewel-x-singleton: do not run on Bionic
The test was failing because it did not explicitly set "os_type" and
"os_version", and teuthology defaulted to Ubuntu Bionic.

Jewel is not supported on Ubuntu Bionic, so we should not try to run tests on
that Ubuntu version.

NOTE: This commit is not cherry-picked from master because the fix is
luminous-specific (there is no "rados/upgrade/jewel-x-singleton" test after
luminous).

Signed-off-by: Nathan Cutler <ncutler@suse.com>