osd: set affinity for *all* threads
It turns out schedule_setaffinity(2) only sets the current *thread*'s
affinity, not the entire process.  Look in /proc to identify all of our
children and map those, too.

Note that this doesn't seem to affect the numa_preferred_nid value in
/proc/$pid/[task/$tid/]sched... at least not immediately.  That value does
seem to change for some threads after some load is applied.

Fixes: https://tracker.ceph.com/issues/42054
Signed-off-by: Sage Weil <sage@redhat.com>