osd/PeeringState: base lease support checks on features, not require_osd_release
The require_osd_release value will change at an arbitrary point in time,
midway through a peering interval.  That can leave us in a state where
we are midway through an interval and suddenly enforcing checks on leases
we don't have.

In contrast, the upacting features change at an interval boundary, which
lets us adjust our behavior for the whole interval.

Signed-off-by: Sage Weil <sage@redhat.com>