ceph-backport.sh: add deprecation warning
The script is not maintained in the stable versions of Ceph.

This commit is not a cherry-pick from master, because the script is maintained
in master. It is unmaintained only in the stable versions, like "nautilus".

Also add a --version option which, in this case, will also indicate to the
user that this is an outdated, deprecated version of the script.

Fixes: https://tracker.ceph.com/issues/42194
Signed-off-by: Nathan Cutler <ncutler@suse.com>