ceph-backport.sh: add deprecation warning
The script is not maintained in the stable versions of Ceph.

This commit is not a cherry-pick from master, because the script is maintained
in master. It is unmaintained only in the stable versions, like
"octopus".

Fixes: https://tracker.ceph.com/issues/42194
Signed-off-by: Nathan Cutler <ncutler@suse.com>
(manual cherry pick of commit 365a93bd3fe62369d33d07993dd246e9b4dd63fd)

Conflicts:
    src/script/ceph-backport.sh