rgw: don't parse the rest of request if invalid
Currently for requests that contain data like PUT, even though we send a HTTP
status to the client we still read the data from the socket, which is set to
keep-alive, so we're effectively reading the next request if the client chooses
to send this and hence reading invalid data. This should fix a few cases where
requests after a expected failure continue to fail as beast would throw an
invalid method.

This is done via returning a new error code which allows for frontends to decide
on whether to read further data or not, for eg. in requests with 100-continue
headers, the client wouldn't have sent the body, so for these requests reading
further data on the socket would make the reads possibly the next requests,
whereas for PUT requests with a non 100-continue header, we'd have to read and
discard the rest of the data

Fixes: https://tracker.ceph.com/issues/42208
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>